@@ -3,11 +3,11 @@ description: >2-
   Package for creating canvas animations based on time and math functions.
 
   Fun + canvas -> funvas :)
-version: 0.0.3+1
+version: 0.1.0-nullsafety.0
 repository: https://github.com/creativecreatorormaybenot/funvas
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
@@ -20,4 +20,4 @@ dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  pedantic: ^1.9.2
+  pedantic: ^1.10.0-nullsafety.3
